Implement basic screen functionality
