+++ src/tetris/ScreenType.java
@@ -1,0 +1,8 @@
+package tetris;
+
+/**
+ * Created by max on 2016-04-01.
+ */
+public enum ScreenType {
+    LOADING, MAIN_MENU, GAME, SETTINGS;
+}
