+++ src/tetris/Tetris.java
@@ -1,7 +1,10 @@
 package tetris;
 
 import java.awt.*;
+import java.awt.event.KeyEvent;
+import java.awt.event.KeyListener;
 import java.io.*;
+import java.util.EnumMap;
 
 import javax.swing.JFrame;
 import javax.swing.JPanel;
@@ -9,44 +12,21 @@ import javax.swing.JPanel;
 /**
  * Create and control the game Tetris.
  */
-public class Tetris extends JPanel {
+public class Tetris extends JPanel implements KeyListener {
 
-    private Game game;
-
-    private TextField score, topScore, level, goal;
-    private TetriminoField nextPiece, hold;
-
-    private Modal gameOver, paused;
-
-    // The number of ticks per second
-    public static double START_SPEED = 1.0;
-
-    // The multiplier amount by which the speed increases every level
-    public static double SPEED_GROWTH_FACTOR = 1.2;
+    private Screen currentScreen;
+    private EnumMap<ScreenType, Screen> screens;
 
     /**
      * Sets up the parts for the Tetris game, display and user control.
      */
     public Tetris() {
 
-        // Create the game
-        game = new Game(this);
-
-        // Information Fields
-        score = new TextField("SCORE", 60);
-        topScore = new TextField("TOP SCORE", 60);
-        level = new TextField("LEVEL", 60);
-        goal = new TextField("GOAL", 60);
-        nextPiece = new TetriminoField("NEXT");
-        hold = new TetriminoField("HOLD");
-
-        // Modals
-        gameOver = new Modal("G A M E   O V E R");
-        gameOver.setBody("press any key to play again",
-                "or press q to quit the game");
-        paused = new Modal("P A U S E D");
-        paused.setBody("press escape to resume play",
-                "or press q to quit the game");
+        // Create screens
+        screens = new EnumMap<>(ScreenType.class);
+        // screens.put(ScreenType.MAIN_MENU, new MenuScreen(this));
+        screens.put(ScreenType.GAME, new GameScreen(this));
+        setScreen(ScreenType.GAME);
 
         // Create window
         JFrame f = new JFrame("Tetris");
@@ -57,25 +37,30 @@ public class Tetris extends JPanel {
         f.setResizable(false);
         f.pack();
         f.setVisible(true);
-        EventController ec = new EventController(game);
-        f.addKeyListener(ec);
+        f.addKeyListener(this);
         setBackground(ColorScheme.BASE_02.color);
     }
 
     /**
-     * Updates the display.
+     * Sets the current screen to the one of the specified type.
+     *
+     * @param s the screen type to switch to
+     */
+    public void setScreen(ScreenType s) {
+        Screen screen = screens.get(s);
+
+        if (screen != null) {
+            if (currentScreen != null) currentScreen.unload();
+            currentScreen = screen;
+            currentScreen.load();
+        }
+    }
+
+    /**
+     * Updates the current screen.
      */
     public void update() {
-        if (game != null) {
-            if (game.isOver()) {
-//                gameOver.setBody("you scored 10 points");
-                gameOver.setVisible(true);
-            } else {
-                gameOver.setVisible(false);
-            }
-
-            paused.setVisible(game.isPaused());
-        }
+        currentScreen.update();
         repaint();
     }
 
@@ -87,52 +72,35 @@ public class Tetris extends JPanel {
     public void paintComponent(Graphics g) {
         super.paintComponent(g);
 
-        // Draw the actual game
-        game.draw(g);
-
-        // Convert Graphics into Graphics2D for better rendering effects
-        Graphics2D g2d = (Graphics2D) g;
-
-        // Turn on anti-aliasing
-        g2d.setRenderingHint(
-                RenderingHints.KEY_ANTIALIASING,
-                RenderingHints.VALUE_ANTIALIAS_ON);
-
-        // Calculate sidebar offsets
-        int rightSide = getWidth() / 2 + 130;
-        int leftSide = getWidth() / 2 - 230;
-
-        // Draw score
-        score.setValue(Integer.toString(game.getScore()));
-        score.draw(g2d, rightSide, 420 - 160);
-
-        // Draw top score
-        topScore.setValue(Integer.toString(game.getHighScore()));
-        topScore.draw(g2d, rightSide, 420 - 80);
-
-        // Draw level
-        level.setValue(Integer.toString(game.getLevel()));
-        level.draw(g2d, leftSide, 420 - 160);
-
-        // Draw goal
-        goal.setValue(Integer.toString(game.getGoal()));
-        goal.draw(g2d, leftSide, 420 - 80);
-
-        // Draw nextPiece
-        nextPiece.setTetrimino(game.getNextPiece());
-        nextPiece.draw(g2d, rightSide, 20);
-
-        // Draw hold
-        hold.setTetrimino(game.getHeldPiece());
-        hold.draw(g2d, leftSide, 20);
-
-        // Draw Game Over modal
-        if (gameOver.isVisible()) gameOver.draw(g2d, getWidth(), getHeight());
-
-        // Draw pause modal
-        if (paused.isVisible()) paused.draw(g2d, getWidth(), getHeight());
+        currentScreen.draw(g);
     }
 
+    /**
+     * Delegates key presses to the current screen.
+     *
+     * @param e the KeyEvent containing info about the key pressed
+     */
+    public void keyPressed(KeyEvent e) {
+        // If key pressed was 'Q', quit the game
+        if (e.getKeyCode() == KeyEvent.VK_Q) {
+            // timer.stop();
+            AudioManager.stopAll();
+            ((JFrame) e.getSource()).dispose();
+            System.exit(0);
+            return;
+        }
+
+        // Send to current screen
+        currentScreen.keyPressed(e);
+    }
+
+    // We don't use these but must implement them to satisfy the interface
+    public void keyReleased(KeyEvent e) {}
+    public void keyTyped(KeyEvent e) {}
+
+    /**
+     * Loads all application resources.
+     */
     private static void loadResources() {
         AudioManager.init();
